Both the cons are duplicates of 'Last One Only Breadcrumbs'
Raised in issue https://github.com/bradfrost/this-is-responsive/issues/31 .

There are cons for this one, but I wasn't sure how to word it.

This doesn't seem like an obvious or familiar interaction. This could be why there are no examples from the wild. It's a very interesting pattern though. I've got a similar challenge on a project and will try some things like this.